requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-tailwind==3.7.0
pillow==10.1.0
python-dotenv==1.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
gunicorn==21.2.0

# Development dependencies
black==23.11.0
isort==5.12.0